### 2025-03-18 #7924 @bastienqueste
We also had an issue pushing to PyPI so the relevant version is actually v0.0.10 rather than v0.0.9 now.
