### 2025-05-06 #7924 @rwegener2
👋🏻 @nbronikowski or @truedichotomy, would you be willing to review this submission for JOSS? [JOSS](https://joss.theoj.org/) is an open source journal for scientific research software. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
