### 2025-05-12 #7924 @nbronikowski
Yes - how much time do I have?
