### 2025-06-04 #8342 @editorialbot
[REVIEW]: gliderad2cp: A Python package to process Nortek AD2CP velocity profiles from gliders
**Submitting author:** @bastienqueste (<a href="http://orcid.org/0000-0002-3786-2275">Bastien Queste</a>)
**Repository:** https://github.com/bastienqueste/gliderad2cp
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.9
