### 2025-06-04 #8342 @rwegener2
👋🏼 @bastienqueste @truedichotomy, @nbronikowski this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
