### 2025-06-13 bastienqueste/gliderad2cp#79 @nbronikowski
@openjournals/joss-reviews#8342 - Comments
@openjournals/joss-reviews#8342

Hi @bastienqueste @callumrollo - thank you so much for producing this software package! I have had no issues running your code with the supplied example to verify the functionality.

I have tried to then adopt your example for your a Slocum glider mounted AD2CP deployment - one I have been looking at previously w. Joe Gradone and Clark Richards using various packages (Python and Matlab). My thoughts so far:

- I had to fix the interpolation function to be able to handle sparse data or some gaps in the data (process_shear.py and the tools.py code) - not a big deal. I will upload to my own fork.
- It would be nice to have some of the code interacting with glider data generalized, so that for example the Slocum glider depth-averaged currents could be imported directly - I haven't been able to get your code block working for Slocum glider data (still trying to get it work).
