### 2025-06-13 bastienqueste/gliderad2cp#79 @bastienqueste
Interesting to hear, are you happy to share that data? As we've tested on slocum data, and have also validated on interrupted datasets, I'm surprised to hear it.

As for the DAC, the capability to implement DAC inputs is not there at the moment but is a planned development (only very little to change). It was carefully and deliberately decided to stick to a DVL approach due to its generally greater accuracy compared to DAC and making it compatible across all gliders including when engineering data is absent.
