### 2025-06-13 bastienqueste/gliderad2cp#79 @nbronikowski
Hi @bastienqueste, yes off course! I have sent you a link via email to a google drive folder with all the data in it. Let me know if you have any issues.
