### 2025-06-20 #8342 @rwegener2
Thanks for attending to your review so quickly @nbronikowski! Did you have any feedback for the author? The JOSS review process is a great place to suggest changes and help submitting authors improve their packages. If you made suggestions as Github issues that's great. In that case its helpful if you can mention `openjournals/joss-reviews#8342` so they get tagged in this issue and I can track them. Thanks!
