### 2025-09-22 #8342 @bastienqueste
@rwegener2 @truedichotomy - can we move forward with this review please?
