+++ gliderad2cp/process_compass.py
@@ -1,0 +1,278 @@
+import numpy as np
+# import pandas as pd
+# import xarray as xr
+import matplotlib.pyplot as plt
+import cmocean.cm as cmo
+
+from scipy.optimize import fmin
+
+from .tools import plog
+rmsd = lambda x,y : np.sqrt(np.nanmean( (x-y)**2 ))
+
+def add_declination(ADCP):
+
+    return ADCP
+
+def correct_heading(ADCP, options):    
+    ## STEP 1 : collect magnetic target
+    # Technically not that big a deal unless covering multiple different areas, in which case we should change the fielf over time, which we don't. Basically, any old constant will do.
+    # Get local geomagnetic target strength:
+#     def getGeoMagStrength():    
+#         lat = np.nanmedian(ADCP.latitude.values)
+#         lon = np.nanmedian(ADCP.longitude.values)
+#         date = np.nanmedian(ADCP.time.values)
+#         year = date.year
+#         month = date.month
+#         day = date.day
+
+#         url = str('http://geomag.bgs.ac.uk/web_service/GMModels/igrf/13/?'+
+#               'latitude='+str(lat)+'&longitude='+str(lon)+
+#               '&date='+str(year)+'-'+str(month)+'-'+str(day)+
+#               '&resultFormat=csv')
+#         import urllib
+#         magdata = urllib.request.urlopen(url)
+
+#         string = 'empty'
+#         while not not string:
+#             out = magdata.readline().decode("utf-8")
+#             if 'total-intensity units="nT"' in out:
+#                 string = out
+#                 print(out)
+#             break
+#     target = float(string.split('>')[1].split('<')[0])
+#     nT2milligauss = 10**-9 * 10000 * 1000 # To tesla, then to gauss then to millgauss
+#     print('Target = '+str(target*nT2milligauss))
+#     return target*nT2milligauss
+    
+    target = 500 #getGeoMagStrength()
+    
+    
+    ## STEP 2 : Extract ADCP magnetometer data
+    if options['ADCP_mounting_direction'] == 'top':
+        sign = -1
+    elif options['ADCP_mounting_direction'] == 'bottom':
+        sign = 1
+    
+    MagX = ADCP['MagnetometerX'].values
+    MagY = ADCP['MagnetometerY'].values
+    MagZ = ADCP['MagnetometerZ'].values
+    
+    AccX = ADCP['AccelerometerX'].values
+    AccY = ADCP['AccelerometerY'].values
+    AccZ = ADCP['AccelerometerZ'].values
+        
+    roll = ADCP['Roll'].values
+    pitch = ADCP['Pitch'].values
+    
+    
+    ## STEP 3 : define lambda functions
+    rmsd    = lambda x,y,z : np.sqrt( np.mean( ( norm(x[idx],y[idx],z[idx]) - target)**2 ) ) # Score function for Hard iron # Notice the [idx] here which removes bad points from the score function!
+    rvar    = lambda x,y,z : np.nanvar(norm(x[idx],y[idx],z[idx])) # Score function for soft iron # Notice the [idx] here which removes bad points from the score function!
+    
+    norm    = lambda x,y,z : np.sqrt(x**2 + y**2 + z**2) 
+
+    def circ(x):
+        x[x < -180] = x[x < -180] + 360
+        x[x >  180] = x[x >  180] - 360
+        return x
+    
+    cosd    = lambda x : np.cos(np.deg2rad(x))
[…]
