### 2025-09-23 b87f760
Start of least squared inversion code development
Initial commit laying out options for choice of profile constructing function (shear vs lsq inversion), as well as options for lsq_inversion_smoothing and lsq_inversion_weighting.

Space for adding the lsq_inversion code is in process_currents under the def of "_lsq_inversion".
