### 2025-09-23 d48ae4a
And now streamlined the old shear integration method.
