### 2025-09-29 #8342 @nbronikowski
> Hi [@bastienqueste](https://github.com/bastienqueste), thanks for the bump. I want to sincerely apologize for not pushing this review forward more actively. I’ve had some large events in my personal life the last few months that have kept me from maintaining the forward momentum on this review, but I’m focused again and I’d like to keep pushing things forward.
>
> Due to the lack of response over the past months I will remove [@truedichotomy](https://github.com/truedichotomy) as a reviewer and search for another reviewer.
>
> In the meantime, [@nbronikowski](https://github.com/nbronikowski) can you please provide some more detail about your review? A one paragraph summary is helpful for myself and the package authors. The JOSS review is designed to be an open place to suggest feedback to package authors, thus helping improve the overall quality of scientific software. It is uncommon that a reviewer doesn’t have any feedback. Are there any small suggestions you considered while reviewing the package?

@rwegener2 - yes, I tried to provide some feedback on the package and my experience using it here: https://github.com/bastienqueste/gliderad2cp/issues/79#issue-3143999641

I am think I provided some good feedback. I also made some changes to the code in the forked repo on my github profile. I have actually since tried another different dataset and made more minor modifications to the code and look forward to a chance to chat with the authors at some point. But I think for the purpose of this review the feedback I previously provided is sufficient. Overall I think a great contribution!
