### 2025-09-30 #8342 @wingtorres
hi @rwegener2 thanks for the tag! I haven't dealt with glider data myself before, but I have some colleagues who are more familiar with that data processing pipeline. Let me ask around at our group meeting this Thursday and I'll get back to you.
