### 2026-02-22 #8342 @JGradone
@rwegener2 I have added my comments as issues in @bastienqueste's repo, but I cannot complete the reviewer checklist because the @editorialbot says I am not a reviewer.

I would check all of the boxes and recommend this awesome package once the points I documented are resolved!
