### 2026-03-05 #8342 @rwegener2
Thanks so much @JGradone and @nbronikowski for your reviews! I really appreciate you both taking the time to contribute to JOSS.

@bastienqueste at this point I have a series of checks that I will go through as editor, before passing this submission on to the Track editor for final review.
