### 2026-03-09 #8342 @rwegener2
Hi @bastienqueste,

I have a few pieces of feedback on your paper. It has a lot of good content, but it could use some proofreading and I’d like to ask you to rearrange the content slightly to align with the JOSS paper structure.

### Proof Reading
- line 20: I believe that “signal-to-noise” should have hyphens
- line 25: There is either an error or a rendering problem with “30\~m” and “15\~m”. I suggest removing ~ altogether and simply using the word “approximately” in both cases
- Figure 2 caption: “The purpose …” The purpose of what?
- line 35: “The toolbox collects successive measurements …” Does this package collect data? Or does it open data collected from the ADCP / from the glider?
- line 50-52: sentence is difficult to follow. Short clause at the end (“…, but lacking an absolute reference”) is not clear
- line 55-57: Is this a complete sentence? It is difficult to follow.
- line 58-62: This is a very long run-on sentence. I suggest breaking it into two sentences for clarity.
- line 65: “none of are developed” Is there a word missing here?
- line 66-69: Is there a word missing at the beginning of the sentence "Individual toolboxes for specific gliders or linked to published papers…”?

### Structure
- Summary: The Summary should be a short description of what your software accomplishes at a high level. It should be about one long paragraph or two shorter paragraphs in length. After reading your paper I imagine your summary would frame the fact that 1) ADCP and Gliders are both useful instruments and that sometimes ADCP data is collected on a glider 2) It is difficult to align the measurements from the two instruments 3) There are lots of types of gliders and, before your tool, no single package can align ADCP and glider measurements from such a wide array of Glider platforms. You might also highlight in 1-2 sentences the many processing steps that your package completes.
- Statement of Need: I think your statement of need section could contain the first two paragraphs of your current Summary plus your current Statement of Need section.
- The remaining section with the itemized list of processing steps could go in its own final section. Authors often choose to call this “Key Features” or “Overview of Package Contents"

Here are a couple of JOSS papers that I think are great examples: https://joss.theoj.org/papers/10.21105/joss.07927 or https://joss.theoj.org/papers/10.21105/joss.08111

JOSS has recently updated our submission guidelines with a new AI policy so the docs are unfortunately not as clear right now about the paper format as they could be. If you have any questions, though, please don't hesitate to ask.
