### 2026-04-27 #8342 @rwegener2
Hey @bastienqueste 👋🏻 Any estimate as to when you'll be able to respond to paper comments?
